refactor(web): dissolve 4 pastas features/*/components/redesign/ + 6 renames base-morta (038 Slice B2)
- consultation: deleta ConsultationSections legacy (dead) + flatten/rename ConsultationView+Sections+Adherence
- medications: flatten (incl. subpasta medicines/) + MedicineCardRedesign->MedicineCard
- protocols: flatten 12 componentes (ProtocolRow, TreatmentTabBar, AnvisaSearchBar, etc)
- stock: flatten + StockCardRedesign->StockCard, CriticalAlertBanner, EntradaHistorico
- dashboard: InsightCardRedesign->InsightCard, SmartAlertsRedesign->SmartAlerts
- protocols: ReminderSuggestionRedesign->ReminderSuggestion
- aliases @medications/@stock/@protocols/components/redesign/* -> sem /redesign/
- AP-239 (colisao flatten vs legacy)

Deferido Slice C: CSS BEM *-redesign__* + views/*Redesign
