+++ .agent/memory/ANTI_PATTERNS_INDEX.md
@@ -3,6 +3,7 @@
 ## 📦 Data & Schema (`data_and_schema`)
 - **[AP-237]** Endpoint de LLM (`api/chatbot.js`) **sem auth** + `systemPrompt` montado no **cliente** → abuso anônimo de quota Groq + bypass das "REGRAS ABSOLUTAS" (cliente removia guardrails SaMD). Fix: auth Supabase JWT obrigatória (401), system prompt composto NO SERVIDOR (cliente só envia `patientContext`), safety guard + rate limit server-side. Validação só client-side é UX, não segurança -> [`anti-patterns/infra_and_deploy/AP-237.md`](./anti-patterns/infra_and_deploy/AP-237.md)
 - **[AP-238]** Path alias vive em 2 configs (`vite.config.js` + `eslint.config.js`); mover destino de alias atualizando só o vite deixa `import-x/no-unresolved` no lint (resolver eslint é separado). Build/test resolvem pelo vite e mascaram. Fix: ao mover pasta com alias, grep+atualizar AMBOS; `rtk lint` é o gate que pega. Detectado 038 Slice A (`@settings`) -> [`anti-patterns/infra_and_deploy/AP-238.md`](./anti-patterns/infra_and_deploy/AP-238.md)
+- **[AP-239]** Flatten de pasta `redesign/` → `components/` colide com arquivo legacy de mesmo nome no destino (não só a base homônima do componente principal). Premissa "base morta deletada no B1 = collision-free" é insuficiente. Fix: antes do `git mv`, `ls` destino + cruzar TODOS os nomes-alvo; colisão com legacy dead (0 importers) → `git rm` antes; build checkpoint por pasta. Detectado 038 Slice B2 (`ConsultationSections` legacy) -> [`anti-patterns/react_and_ui/AP-239.md`](./anti-patterns/react_and_ui/AP-239.md)
 - **[AP-236]** Ler metadado de ancoragem (`instance_id`) do objeto **pós-Zod** → `undefined` (não é coluna, `logSchema` o estrita no `safeParse`) → ancoragem em lote quebrada (toda dose vira avulsa). Bateu web E mobile no mesmo PR #673. Fix: mapear sobre `validatedLogs` mas ler o metadado do **input cru por índice** (`logsData[index].instance_id`). Irmão de AP-214 (sintoma igual, fix oposto: lá adiciona ao schema, aqui nunca espera do schema). Detectado review Gemini #673 -> [`anti-patterns/data_and_schema/AP-236.md`](./anti-patterns/data_and_schema/AP-236.md)
 - **[AP-231]** Mutar `medicine_logs` direto (UPDATE/DELETE) sem replay dos efeitos colaterais do registro (estoque FIFO + âncora da instância) → corrupção. Dois casos no PR #668: (a) **ghost taken** web — `logService.delete` deletava o log e a FK `ON DELETE SET NULL` limpava `medicine_log_id` mas deixava `dose_instance.status='taken'` → dose fantasma contando na adesão; (b) **furo de estoque** mobile — `useHistoryMutation` mutava log via cliente bypassando o FIFO. Fix: toda mutação no service (`logService`/`doseService.updateOrphanLog`/`deleteOrphanLog`) restaura+reconsome estoque e reverte instância; paridade web↔mobile↔bot. FK SET NULL nunca reverte campos de estado dependentes. Irmão de AP-193/AP-185. Detectado smoke+review #668 -> [`anti-patterns/data_and_schema/AP-231.md`](./anti-patterns/data_and_schema/AP-231.md)
 - **[AP-228]** Prefill com `''` em campo `z.enum().nullable()` (ex.: `intake_unit ?? ''`) → `''` não é membro do enum NEM `null` → update falha mudo. Pior quando o input do campo é condicional (renderiza só p/ líquidos): erro "campo obrigatório" sem campo visível. Caso real: editar tratamento não-líquido (smoke C). Fix: fallback `?? null`. Smoke deve cobrir edit de TODA variante (líquido E sólido). Irmão de AP-225/AP-214. Detectado smoke #665 -> [`anti-patterns/data_and_schema/AP-228.md`](./anti-patterns/data_and_schema/AP-228.md)
