+++ .agent/memory/anti-patterns/react_and_ui/AP-239.md
@@ -1,0 +1,18 @@
+# AP-239 — Flatten de pasta `redesign/` colide com arquivo legacy de mesmo nome no destino
+
+**Categoria:** react_and_ui · **Detectado:** 038 Slice B2 (2026-06-23)
+
+## Sintoma
+Ao achatar `features/consultation/components/redesign/` → `components/`, o `git mv ConsultationRedesignSections.jsx → ConsultationSections.jsx` falhou com `fatal: destination exists`. Já existia um `components/ConsultationSections.jsx` **legacy** (versão antiga, emoji/classes diferentes, exports `ConsultationMedicinesSection`). A sed de identificadores rodou mesmo assim e cruzou imports entre o arquivo movido e o legacy, quebrando os named exports.
+
+## Causa
+Premissa errada: "a base `*Redesign` tinha base morta deletada no B1, logo o rename é collision-free". Mas o flatten de uma pasta inteira traz arquivos cujos nomes-alvo podem colidir com **outros** arquivos legacy no diretório destino — não só com a base homônima do componente principal. `ConsultationSections` (legacy dead, 0 importers) não era a base do `ConsultationView`, mas colidia com o `RedesignSections` ao normalizar o nome.
+
+## Fix / Prevenção
+Antes de achatar `redesign/` → `components/`, para CADA arquivo que será movido/renomeado:
+1. `ls` o diretório destino e cruzar nomes-alvo contra os existentes (não só a base do componente principal).
+2. Para cada colisão, `grep -rln <symbol>` os importers do legacy. Se **0 importers** → `git rm` o legacy ANTES do `git mv` (dead code, AP-164).
+3. Só então `git mv` + blanket replace de identificadores.
+4. Build checkpoint por pasta (não só no fim) — pega import quebrado cedo.
+
+Relacionado: [[AP-164]] (barrel/dead code), AP-H27 (git mv quebra import), AP-238 (alias em 2 configs).
