+++ .agent/state.json
@@ -8,7 +8,7 @@
     "ceremonies_run": ["eng-review"],
     "ceremony_findings_count": 3,
     "ceremony_scope_decisions": ["038 Slice A entregue (PR #677)", "Slice B fatiado em B1 (dead code + 4 renames collision-free) e B2 (dissolver redesign/ + 6 renames base-morta)", "B1 entregue local — aguarda PR", "guard FULL"],
-    "goal": "038 Slice B1 entregue (dead code + 4 renames collision-free) — abrir PR; B2/C pendentes",
+    "goal": "038 Slice B2 entregue (dissolver 4 pastas redesign/ + 6 renames base-morta) — abrir PR; Slice C pendente (deriveProtocolStatus carve-out + views/*Redesign + CSS BEM classes)",
     "goal_type": "refactor",
     "tier": 1,
     "spec_dir": "plans/specs/038-refactor-web-structure",
@@ -23,9 +23,9 @@
   },
   "memory": {
     "last_distillation": "2026-06-19T21:01:22.841Z",
-    "journal_entries_since_distillation": 9,
+    "journal_entries_since_distillation": 10,
     "rules_count": 228,
-    "anti_patterns_count": 235,
+    "anti_patterns_count": 236,
     "decisions_count": 72,
     "contracts_count": 25,
     "distillation_due": false
