+++ CHANGELOG.md
@@ -84,6 +84,7 @@ Cuidando da sua rotina com carinho e zero complicação. 💙
 
 - **Spec 038 / Slice A:** o redesign da experiência do paciente foi entregue, mas a pasta de produção das views ainda se chamava `views/redesign/`. `git mv views/redesign/* → views/` (Dashboard, Stock, Treatments, Medicines, HealthHistory, Settings, Emergency, Profile, Consultation, NotificationInbox + subpastas `history/ profile/ settings/`); artefatos Landing agrupados em `views/landing/`. Atualizados `AppViewRouter` (12 lazy imports), `vite.config.js` (alias `@settings` + manualChunks `HealthHistory`/`Stock`/`Landing`), o resolver `@settings` do `eslint.config.js` (alias duplicado fora do vite — AP-238) e 3 arquivos de teste. Refator puro, sem mudança de comportamento. Slices B (dissolver `*Redesign`/dead code) e C (carve-out + doc) a seguir.
 - **Spec 038 / Slice B1:** removido dead code legado (7 componentes pré-redesign com 0 importers: `InsightCard`, `SmartAlerts`, `StockCard`, `MedicineCard`, `ConsultationView`, `ReminderSuggestion` + a base `PrescriptionTimeline` com test/css — testava código morto) e renomeados 4 componentes `*Redesign` collision-free para a base (`RingGaugeRedesign`→`RingGauge`, `CostSummaryRedesign`→`CostSummary`, `PrescriptionTimelineRedesign`→`PrescriptionTimeline`, `BottomNavRedesign`→`BottomNav` — arquivo + css + identificador + importers). −2843 linhas. Refator puro, sem mudança de comportamento. B2 (dissolver pastas `redesign/` + 6 renames com base morta) e C (carve-out + doc) a seguir.
+- **Spec 038 / Slice B2:** dissolvidas as 4 pastas `features/{consultation,medications,protocols,stock}/components/redesign/` — conteúdo achatado para `components/` (incl. subpasta aninhada `medicines/` de medications). Renomeados 6 componentes `*Redesign` cujas bases pré-redesign ainda existiam: removida a base morta colidente (`ConsultationSections` legado, 0 importers) e promovido o `*Redesign` para o nome base (`InsightCardRedesign`→`InsightCard`, `SmartAlertsRedesign`→`SmartAlerts`, `ReminderSuggestionRedesign`→`ReminderSuggestion`, `MedicineCardRedesign`→`MedicineCard`, `ConsultationViewRedesign`→`ConsultationView`, `StockCardRedesign`→`StockCard` + exports internos `Redesign*Section`→`Consultation*Section`). Imports/aliases (`@medications`/`@stock`/`@protocols/components/redesign/*` → sem `/redesign/`) e importers das views atualizados. Pasta `redesign/` extinta em `features/`. **Deferido p/ Slice C:** classes CSS BEM `*-redesign__*` e arquivos `*Redesign` ainda em `views/{settings,emergency,profile,history}/`. Refator puro, sem mudança de comportamento.
 
 ### 🔒 Segurança / ✨ Melhorado (Chatbot IA — endurecimento) — web `4.10.0`→**`4.11.0`** · backend/infra · Telegram
