+++ apps/web/src/features/consultation/components/ConsultationAdherenceSection.jsx
