+++ apps/web/src/features/consultation/components/ConsultationView.css
@@ -1,5 +1,5 @@
 /* ============================================
-   ConsultationViewRedesign — Santuário Terapêutico
+   ConsultationView — Santuário Terapêutico
    ============================================ */
 
 .sr-consultation {
