+++ apps/web/src/features/dashboard/components/InsightCard.css
