+++ apps/web/src/features/dashboard/components/SmartAlerts.css
