+++ apps/web/src/features/medications/components/MedicineCard.css
@@ -1,5 +1,5 @@
 /* ================================================
-   MedicineCardRedesign — Santuário Terapêutico
+   MedicineCard — Santuário Terapêutico
    ================================================ */
 
 .sr-medicine-card {
