+++ apps/web/src/features/medications/components/Medicines.css
@@ -1,5 +1,5 @@
 /* ============================================
-   MedicinesRedesign — Santuário Terapêutico
+   Medicines — Santuário Terapêutico
    ============================================ */
 
 .sr-medicines {
