+++ apps/web/src/features/protocols/components/AdherenceBar7d.jsx
