+++ apps/web/src/features/protocols/components/AdherenceLabel.jsx
