+++ apps/web/src/features/protocols/components/StockPill.jsx
