+++ apps/web/src/features/protocols/components/TreatmentTabBar.jsx
