+++ apps/web/src/features/stock/components/CriticalAlertBanner.jsx
