+++ apps/web/src/features/stock/components/EntradaHistorico.jsx
