+++ apps/web/src/features/stock/components/StockCard.css
@@ -1,5 +1,5 @@
 /* ============================================
-   StockCardRedesign — Card de estoque
+   StockCard — Card de estoque
    REGRA CARDINAL: Zero hex codes. Apenas var(--color-*).
    ============================================ */
