+++ apps/web/src/views/TreatmentsComplex.jsx
@@ -9,8 +9,8 @@ import { useState, useMemo } from 'react'
 import { motion } from 'framer-motion'
 import { useMotion } from '@shared/hooks/useMotion'
 import { useDashboard } from '@dashboard/hooks/useDashboardContext'
-import TreatmentPlanHeader from '@protocols/components/redesign/TreatmentPlanHeader'
-import ProtocolRow from '@protocols/components/redesign/ProtocolRow'
+import TreatmentPlanHeader from '@protocols/components/TreatmentPlanHeader'
+import ProtocolRow from '@protocols/components/ProtocolRow'
 import ProtocolRiskBadge from '@adherence/components/ProtocolRiskBadge'
 import { calculateProtocolRisk } from '@adherence/services/protocolRiskService'
